Broke out EmailForm into own component
