+++ src/constants/email-text-field-type.ts
@@ -1,4 +1,9 @@
-export const to = 'to';
-export const from = 'from';
-export const subject = 'subject';
-export const body = 'body';
+export default class EmailTextFieldTypeConstants {
+  static readonly TO: string = 'to';
+
+  static readonly FROM: string = 'from';
+
+  static readonly SUBJECT: string = 'subject';
+
+  static readonly BODY: string = 'body';
+}
